nvidia_vad_test/requirements.txt
# fastapi
# uvicorn
python-multipart
# numpy

# mlserver
torchaudio
soundfile
nemo_toolkit[asr]
resampy
huggingface-hub==0.23.2